# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RuleOutcomeView": {
        "additionalProperties": false,
        "properties": {
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "queue": {
            "title": "Queue",
            "type": "string"
          },
          "segment": {
            "$ref": "#/components/schemas/Segment"
          }
        },
        "required": [
          "segment",
          "priority",
          "queue"
        ],
        "title": "RuleOutcomeView",
        "type": "object"
      },
      "RuleSetView": {
        "additionalProperties": false,
        "properties": {
          "default": {
            "$ref": "#/components/schemas/RuleOutcomeView"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/RuleView"
            },
            "title": "Rules",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "rules",
          "default"
        ],
        "title": "RuleSetView",
        "type": "object"
      },
      "RuleView": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "then": {
            "$ref": "#/components/schemas/RuleOutcomeView"
          },
          "when": {
            "additionalProperties": {
              "type": "object"
            },
            "title": "When",
            "type": "object"
          }
        },
        "required": [
          "id",
          "description",
          "when",
          "then"
        ],
        "title": "RuleView",
        "type": "object"
      },
      "Segment": {
        "enum": [
          "FAST_TRACK",
          "STANDARD",
          "COMPLEX",
          "SPECIAL_INVESTIGATION"
        ],
        "title": "Segment",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "OIDC access token issued by the corporate identity provider.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "claims-platform@meridian-assurance.example",
      "name": "Claims Platform"
    },
    "description": "Claims intake and servicing for Meridian Assurance personal lines.\n\nImplements the requirements in `docs/PRD.md`: first notification of loss, triage against a\nversioned rule set, reserve authority, settlement, documents, SLA clocks and an immutable\naudit trail.",
    "title": "Meridian Assurance Claims Intake API",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/rule-sets/current": {
      "get": {
        "operationId": "current_rule_set_v1_rule_sets_current_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read the active triage rule set",
        "tags": [
          "system",
          "triage"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://claims.meridian-assurance.example"
    },
    {
      "description": "Local",
      "url": "http://localhost:8000"
    }
  ]
}